Base Files
These are the base files required to run the function 'coordtrans.m'. The files names 'flips' and 'rotates' are also functions needed for reflections and rotations around axes. 'Coordtrans' also contains several embeded functions that are needed for it to run properly. The function 'DMS2degrees' has been defined as its analogue 'dms2degrees' is only available for MATLAB 2018b or newer.